회원가입 페이지 및 분석보고서 페이지

+++ src/main/java/com/example/etProject/dto/ProducersDTO.java
@@ -20,7 +20,7 @@ public class ProducersDTO {
     private String memberId;
     private Double locationX;
     private Double locationY;
-    private Double installedCapacity;
+    private int installedCapacity;
     private Boolean isProduce;
